requirements.txt
django==2.2.12
djangorestframework==3.11.0
djoser==2.0.3